requirements.txt
django==3.2
mysqlclient
djangorestframework
django-cors-headers